# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitnes Planer API",
        "description": "Dokumentacija koristeći PHP Atribute",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Lokalni API server"
        }
    ],
    "paths": {
...
        "/vezbaci": {
            "get": {
                "tags": [
                    "Trener"
                ],
                "summary": "Prikazuje vežbače ulogovanog trenera",
                "operationId": "2ebfbba596dc33a0db25bcc8a332d6b0",
                "responses": {
                    "200": {
                        "description": "Lista vežbača"
                    },
                    "403": {
                        "description": "Nije trener"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{user}/parametri": {
            "get": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Prikazuje parametre određenog korisnika",
                "operationId": "a8adc67c3929cf25e818beb0d5c9fff3",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista parametara"
                    }
                }
            },
            "post": {
                "tags": [
                    "Trener"
                ],
                "summary": "Dodavanje novih parametara korisniku (samo trener)",
                "operationId": "53d7d9873e9555a5cad90025fca86826",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "tezina",
                                    "visina"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tezina": {
                                        "type": "number"
                                    },
                                    "visina": {
                                        "type": "number"
                                    },
                                    "bmi": {
                                        "type": "number"
                                    },
                                    "masti": {
                                        "type": "number"
                                    },
                                    "misici": {
                                        "type": "number"
                                    },
                                    "obim_struka": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parametri dodati"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{user}/ciljevi": {
            "get": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Prikazuje ciljeve određenog korisnika",
                "operationId": "afd6d804eed25b20545280c6a4d9c018",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista ciljeva"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ciljevi"
                ],
                "summary": "Trener dodaje cilj za vežbača",
                "operationId": "48b40a7a329b9636e14e9228ed65965b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cilj kreiran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trener/pretraga": {
            "get": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Pretraga trenera po imenu, prezimenu ili emailu",
                "operationId": "889d13f4c2cead786b87497098a382a5",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rezultati pretrage"
                    }
                }
            }
        },
        "/trener/postavi": {
            "post": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Vežbač bira sebi trenera",
                "operationId": "7fcb39dd4cc017fcdf8aca9669b186e5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trener_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trener/profil": {
            "post": {
                "tags": [
                    "Trener"
                ],
                "summary": "Ažuriranje profila za trenere",
                "operationId": "1f1871888dcb67faea9808db9f53dd85",
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trener/ukloni": {
            "post": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Vežbač uklanja trenutnog trenera",
                "operationId": "69ca3f75eecfd255cb4ff0d3b67c186a",
                "responses": {
                    "200": {
                        "description": "Trener uklonjen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{user}": {
            "get": {
                "tags": [
                    "Admin",
                    "Trener"
                ],
                "summary": "Prikaz detalja korisnika sa parametrima i ciljevima",
                "operationId": "a8a7a1f979a2abb839b4b320238106fc",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji korisnika"
                    }
                }
            }
        },
...
        "/admin/complete-users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Kompletna lista svih korisnika (Admin)",
                "operationId": "bbe1cf2b5f615558da4bca66b67ec29f",
                "responses": {
                    "200": {
                        "description": "Svi podaci"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/assign-vezbac-trener": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin povezuje vežbača i trenera",
                "operationId": "d05b128d158a98aebf7f914836df9682",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trener_id": {
                                        "type": "integer"
                                    },
                                    "vezbac_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno povezano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trener/{trener_id}/vezbaci": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Prikaz svih vežbača određenog trenera",
                "operationId": "3afd37d7a4812bf2260fcae47147515e",
                "parameters": [
                    {
                        "name": "trener_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            }
        },
        "/vezbac/{vezbac_id}/trener": {
            "get": {
                "tags": [
                    "Admin",
                    "Korisnik"
                ],
                "summary": "Prikaz trenera određenog vežbača",
                "operationId": "1e083e2b92e34f14deb0390e98b9b2e2",
                "parameters": [
                    {
                        "name": "vezbac_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podaci o treneru"
                    }
                }
            }
        },
        "/korisnik/profil": {
            "post": {
                "tags": [
                    "Korisnik"
                ],
                "summary": "Ažuriranje profila za vežbače",
                "operationId": "8003a3da41d28932b9b42dcc94236959",
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}